style(raijin): rename initializer interface file
