+++ yarn/raijin/src/index.ts
@@ -1,3 +1,3 @@
-export type { RunRaijinInitializerOptions } from './initializer/options.js'
+export type { RunRaijinInitializerOptions } from './initializer/interface.js'
 
 export { runRaijinInitializer } from './initializer/index.js'
